Eliminate compiler warnings
The indentation of the code is misleading as to what is or isn't part of the preceding if's and for's.  

Fix the indentation to match how the compiler will interpret the code.  Would be better to use explicit open/close braces, but this is the minimal change to eliminate compiler warnings.